Fixed bug in test.
Got the following error message while executing the tests:

```
>./test
[Started testing]

[Running: tokenizers]

[Started section: 'strspn']
test.cpp:119: (next == __null) failed for: false
[End of section: 'strspn' 1 of 13 assertions failed]

[Finished: 'tokenizers' 1 test case failed (1 of 13 assertions failed)]

[Testing completed. 1 of 3 test cases failed (1 of 31 assertions failed)]
```

The current pointer in the code was actually moved one position to the
right of the terminal `\0` character of the const char string.
Probably the test passed for others, since the memory after the string
was filled with `\0` bytes.